requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
black==24.4.2
boto3==1.34.106
botocore==1.34.106
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.7
dataclass-type-validator==0.1.2
Deprecated==1.2.14
distlib==0.3.8
Django==4.2.13
django-debug-toolbar==4.3.0
django-filter==24.2
django-sendgrid-v5==1.2.3
django-storages==1.14.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
filelock==3.13.4
idna==3.7
inflection==0.5.1
isort==5.13.2
jmespath==1.0.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kombu==5.3.7
minio==7.2.7
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pipenv==2023.12.1
platformdirs==4.2.2
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
PyGithub==2.3.0
PyJWT==2.8.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
PyYAML==6.0.1
redis==5.0.4
referencing==0.35.1
requests==2.31.0
rpds-py==0.18.1
s3transfer==0.10.1
sendgrid==6.11.0
sentry-sdk==2.2.0
six==1.16.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
stripe==9.6.0
tomli==2.0.1
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
wrapt==1.16.0
